+++ src/components/course-view/CourseView.css.ts
@@ -1,64 +1,19 @@
 import { style } from '@vanilla-extract/css';
 
 import { COLOR } from '@/styles/foundation';
-import { sprinkles } from '@/styles/sprinkle.css';
 
 export const wrapper = style({
-    maxWidth: '429px',
-    padding: '16px 18px',
+    padding: '12px 8px',
+    backgroundColor: COLOR.MonoWhite,
+    overflowY: 'scroll',
 
-    display: 'flex',
-    alignItems: 'center',
-    columnGap: '12px',
-
-    ':hover': {
-        borderRadius: '100px',
-        backgroundColor: `${COLOR.MonoBlack}08`
-    }
-});
-
-export const indexBox = style([
-    sprinkles({ typography: 'Bold16' }),
-    {
-        width: '26px',
-        height: '26px',
-
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-
-        padding: '0 10px',
-
-        boxSizing: 'border-box',
-        textAlign: 'center',
-
-        border: `1px solid ${COLOR.Gray300}`,
-        borderRadius: '4px',
+    "::-webkit-scrollbar": {
+        width: '8px',
     },
-]);
 
-export const courseBox = style({
-    display: 'flex',
-    flexDirection: 'column',
-    rowGap: '4px',
-    flexGrow: 1,
-});
-
-export const courseName = style([sprinkles({ typography: 'ExtraBold18' })]);
-
-export const addressBox = style({
-    display: 'flex',
-    columnGap: '4px',
-    alignItems: 'center',
-    color: COLOR.LablesTertiary,
-});
-
-export const address = style([
-    sprinkles({ typography: 'Regular15' }),
-    { color: COLOR.LablesTertiary },
-]);
-
-export const listIcon = style({
-    margin: '10px',
-    cursor: 'pointer',
+    "::-webkit-scrollbar-thumb": {
+        width: '3px',
+        borderRadius: '4px',
+        backgroundColor: COLOR.Gray300,
+    },
 });
