+++ src/constants/dummy.ts
@@ -16,5 +16,47 @@ export const dummyCourseList = [
         name: '집',
         originName: '우리집',
         address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 1,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 2,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 3,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 4,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 5,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 6,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
+    },
+    {
+        id: 7,
+        name: '집',
+        originName: '우리집',
+        address: '경기도 화성시 동탄중앙로 189 340동 602호',
     }
 ]
